🟢 Add 0069. Sqrt(x) [Easy] (Python)

+++ Easy/0069-Sqrtx/solution.py
@@ -1,10 +1,10 @@
 # ═══════════════════════════════════════════════════════
 #  Problem  : 0069. Sqrt(x)
-#  URL      : https://leetcode.com/problems/sqrtx/
+#  URL      : https://leetcode.com/problems/sqrtx/submissions/2079687409/
 #  Difficulty : Easy
 #  Language : Python
 #  Runtime  : 0 ms
-#  Memory   : 12.4 MB
+#  Memory   : 12.3 MB
 #  Solved   : July 24, 2026
 # ═══════════════════════════════════════════════════════
 
@@ -14,5 +14,5 @@ class Solution(object):
         :type x: int
         :rtype: int
         """
-        return abs(floor(x ** 0.5))
+        return int(abs(floor(x ** 0.5)))
